# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: LupexWallet API
  version: "1.0.0"
  description: >
    REST API однопользовательского приложения учета личных финансов.
    Технические идентификаторы — на английском, см. docs/architecture/ddd-model.md
    для доменной модели и docs/database/schema.md для схемы БД.
  license:
    name: Proprietary — внутренний проект, не публикуется
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /operations:
    get:
      tags: [Operations]
      summary: Список операций (UC-11, UC-12)
      operationId: listOperations
      parameters:
        - name: walletId
          in: query
          schema: { type: string, format: uuid }
        - name: operationTypeId
          in: query
          schema: { type: string, format: uuid }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Operations]
      summary: Создать операцию дохода/расхода/корректировки (UC-11, UC-12, UC-24)
      operationId: createOperation
      description: Для перевода используйте POST /transfers, а не этот эндпоинт (раздел 4).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OperationCreateRequest" }
      responses:
        "201":
          description: Создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /operations/{operationId}:
    parameters:
      - $ref: "#/components/parameters/OperationIdParam"
    get:
      tags: [Operations]
      summary: Получить операцию
      operationId: getOperation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operation" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Operations]
      summary: Редактировать операцию (UC-13) — каскадный пересчет истории (Q4)
      operationId: updateOperation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OperationUpdateRequest" }
      responses:
        "200":
          description: Обновлена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Операция — часть перевода; редактируется/удаляется только через перевод
            целиком (раздел 4 требований).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
    delete:
      tags: [Operations]
      summary: Удалить операцию (UC-14)
      operationId: deleteOperation
      responses:
        "204": { description: Удалена }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Операция — часть перевода (удалите через DELETE /transfers/{id}) либо
            её дата строго раньше сегодняшнего дня — уже отражена в истории баланса,
            отредактируйте вместо удаления (ADR-0002, ADR-0009).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  # ---------- Transfers ----------
  /transfers:
    get:
      tags: [Transfers]
      summary: Список переводов
      operationId: listTransfers
      parameters:
        - name: walletId
          in: query
          schema: { type: string, format: uuid }
          description: Кошелек как источник ИЛИ получатель.
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransferPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Transfers]
      summary: Создать перевод между кошельками одной валюты (UC-16)
      operationId: createTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferCreateRequest" }
      responses:
        "201":
          description: Перевод создан (атомарно с двумя операциями)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transfer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Валюты кошельков не совпадают (раздел 4) либо кошельки совпадают.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /transfers/{transferId}:
    parameters:
      - name: transferId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Transfers]
      summary: Получить перевод
      operationId: getTransfer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transfer" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Transfers]
      summary: Удалить перевод (обе операции атомарно, UC-16)
      operationId: deleteTransfer
      responses:
        "204": { description: Удален }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Дата перевода строго раньше сегодняшнего дня (ADR-0009).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  # ---------- Exchange rates ----------
...
components:
  parameters:
...
    OperationIdParam:
      name: operationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CursorParam:
      name: cursor
      in: query
      description: Непрозрачный курсор для постраничной навигации.
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100, minimum: 1 }
  schemas:
...
    Money:
      type: object
      required: [amount, currencyId]
      properties:
        amount:
          type: string
          format: decimal
          description: Произвольная точность (ADR-0005); передается строкой, чтобы не терять точность в JSON-числах.
          example: "1234.56789012"
        currencyId: { type: string, format: uuid }
        currencyCode: { type: string, readOnly: true, example: "USD" }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    CursorPageMeta:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }

    # ---- Auth ----
...
    Operation:
      type: object
      required: [id, walletId, operationTypeId, amount, operationDate, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        walletId: { type: string, format: uuid }
        operationTypeId: { type: string, format: uuid }
        amount: { $ref: "#/components/schemas/Money" }
        operationDate: { type: string, format: date }
        adjustmentMode:
          type: [string, "null"]
          enum: [Absolute, Delta, null]
          description: Заполнено только когда operationType ссылается на behaviorKind = Adjustment (Q3).
        transferId: { type: [string, "null"], format: uuid, readOnly: true }
        createdAt: { type: string, format: date-time, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
    OperationCreateRequest:
      type: object
      required: [walletId, operationTypeId, amount, operationDate]
      properties:
        walletId: { type: string, format: uuid }
        operationTypeId: { type: string, format: uuid }
        amount: { type: string, format: decimal }
        operationDate: { type: string, format: date }
        adjustmentMode: { type: string, enum: [Absolute, Delta] }
    OperationUpdateRequest:
      type: object
      properties:
        walletId:
          type: string
          format: uuid
          description: >
            Кошелек операции (UC-13, решение пользователя от 2026-09-11). Если отличается
            от текущего WalletId операции — операция переносится на другой кошелек: дельта
            реверсируется на исходном кошельке и заново применяется на новом (с учетом его
            валюты и AccountingStartDate). Недоступно для операций — частей перевода (409).
        operationTypeId: { type: string, format: uuid }
        amount: { type: string, format: decimal }
        operationDate: { type: string, format: date }
        adjustmentMode: { type: string, enum: [Absolute, Delta] }
    OperationPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Operation" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }

    # ---- Transfers ----
    Transfer:
      type: object
      required: [id, sourceWalletId, targetWalletId, sourceOperationId, targetOperationId, amount, transferDate, createdAt]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        sourceWalletId: { type: string, format: uuid }
        targetWalletId: { type: string, format: uuid }
        sourceOperationId: { type: string, format: uuid, readOnly: true }
        targetOperationId: { type: string, format: uuid, readOnly: true }
        amount: { $ref: "#/components/schemas/Money" }
        transferDate: { type: string, format: date }
        createdAt: { type: string, format: date-time, readOnly: true }
    TransferCreateRequest:
      type: object
      required: [sourceWalletId, targetWalletId, amount, transferDate]
      properties:
        sourceWalletId: { type: string, format: uuid }
        targetWalletId: { type: string, format: uuid }
        amount: { type: string, format: decimal }
        transferDate: { type: string, format: date }
    TransferPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Transfer" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }

    # ---- Balance history ----
...
  responses:
    BadRequest:
      description: Некорректные параметры запроса
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Отсутствует или неверен токен доступа
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Запись не найдена
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT